+++ esp-hal/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add Flex / AnyFlex GPIO pin driver (#1659)
+- Add new `DmaError::UnsupportedMemoryRegion` - used memory regions are checked when preparing a transfer now
 
 ### Fixed